refactor: extract LLMBackendClient to eliminate duplicated HTTP transport
MeetingSummaryClient and MeetingChatClient duplicated ~1000 lines of HTTP
logic for the same 6 backends (OpenAI, OpenRouter, ChatGPT/WHAM, Ollama,
LM Studio, Custom LLM). This was the root cause of lmStudio/customLLM
being missed in MeetingChatClient — a bug fixed in the previous commit.

LLMBackendClient.swift owns all transport, response parsing, URL resolution,
and error handling. LLMMessage and LLMBackendError replace the per-client
types. MeetingChatClient (509 → 57 lines) and MeetingSummaryClient
(1293 → 338 lines) are now thin orchestrators that build prompts and
delegate to LLMBackendClient.send(messages:config:timeout:maxTokens:).

MeetingChatMessage is a typealias for LLMMessage so MeetingChatView needs
no changes. MeetingSummaryClient retains forwarding shims for its URL
resolution helpers so MeetingDetailView needs no changes. MuesliController
updates one error type check from MeetingSummaryError to LLMBackendError.
